--- code/deno/README.md
@@ -1,1 +1,0 @@
-#
