--- code/deno/common/mod.ts
@@ -1,1 +1,0 @@
-export * from './std.lib/mod.ts';
