--- code/deno/mod.ts
@@ -1,1 +1,0 @@
-export * from './common/mod.ts';
