+++ code/deno/std.lib.server/src/filesystem/Fs.cp.ts
@@ -1,0 +1,15 @@
+/**
+ * Copy all files in a directory.
+ */
+export async function copyDir(sourceDir: string, targetDir: string) {
+  await Deno.mkdir(targetDir, { recursive: true });
+  for await (const entry of Deno.readDir(sourceDir)) {
+    const srcPath = `${sourceDir}/${entry.name}`;
+    const destPath = `${targetDir}/${entry.name}`;
+    if (entry.isDirectory) {
+      await copyDir(srcPath, destPath);
+    } else if (entry.isFile) {
+      await Deno.copyFile(srcPath, destPath);
+    }
+  }
+}
