+++ code/deno/std.lib/src/-Testing.test.ts
@@ -1,0 +1,7 @@
+import { Expect } from './mod.ts';
+
+Deno.test('Testing', async (test) => {
+  await test.step('Expect.eql', () => {
+    Expect.eql(123, 123);
+  });
+});
