+++ script.deno/test.ts
@@ -1,3 +1,4 @@
-import { Cmd } from '@sys/stdlib';
+import { Cmd } from '@sys/stdlib-server';
 
-await Cmd.sh('cd code/deno', 'deno test -RW');
+await Cmd.sh('cd code/deno/std.lib', 'deno test -RW');
+await Cmd.sh('cd code/deno/std.lib.server', 'deno test -RW');
